Add files via upload
a coffee machine maker, created by different components in an object-oriented scheme. This project consists of coffee_maker which holds the required data for each type of coffee. The menu and the money machine which gets a hold of processing the coins purchased by the user --- on the main file we run the main loop of program and gathering and implementing all the components